Fix Customer and Shopping Cart repositories and the SQL for Shopping cart and Customer.

+++ src/main/java/desbytes/Repositories/ProductRepository.java
@@ -40,7 +40,7 @@ public class ProductRepository {
                 new ProductRowMapper());
     }
 
-    public Product findProductById(int id) {
+    public Product findProductById(double id) {
         QueryReader reader = new QueryReader();
         String content = reader.readQueryFile("product_queries", "get_product.sql");
         return jdbcTemplate.queryForObject(content,
